requirements.txt
google-cloud-aiplatform>=1.36.0
langchain>=0.0.267
langchain-community>=0.0.1
langchain-google-vertexai>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.2
sqlalchemy>=2.0.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0 